requirements.txt
fastapi
opencv-python
deepface
uvicorn
numpy
google-generativeai